Added more core libraries. Also added copyright to top of all files

+++ frameworks/CoreGraphics/CGBase.h
@@ -1,1 +1,9 @@
-typedef float CGFloat;
+// 
+//  CGBase.h
+//  vienna
+//  
+//  Created by Adam Beynon on 2009-05-02.
+//  Copyright 2009 Adam Beynon. All rights reserved.
+// 
+
+typedef double CGFloat;
